+++ bindings/qml/src/QmlZyre.h
@@ -149,7 +149,7 @@ public:
 public slots:
     //  Return the Zyre version for run-time API detection; returns
     //  major * 10000 + minor * 100 + patch, as a single integer.  
-    int version ();
+    uint64_t version ();
 
     //  Self test of this class.
     void test (bool verbose);
